templates/django/requirements.txt
{% if render_features['django'] %}
Django
djangorestframework
{% endif %}
{% if render_features['django-postgresql'] %}
psycopg2
{% endif %}
{% if render_features['django-redis-rabbit'] %}
redis
{% endif %}
{% if render_features['django-jwt'] %}
djangorestframework-jwt
{% endif %}
{% if render_features['django-socialauth'] %}
-e git://github.com/bmpenuelas/django-rest-framework-social-oauth2.git@fix-rev-call#egg=rest_framework_social_oauth2
{% endif %}
{% if render_features['django-celery'] %}
celery
{% endif %}
{% if render_features['django-gunicorn'] %}
gunicorn
{% endif %}
